requirements.txt
fastapi
uvicorn[standard]
chromadb
sentence-transformers
openai
ollama
tavily-python
httpx
python-dotenv
pytest
pytest-asyncio
httpx
